chore: refactor into subcomponents
